Where to start added mana display to the ui, added a ui for health display too. added a global Cooldown on all spell casts. made BoomerangProjectile draw semi nicely, now i need to make it draw as i want it. removed resources from the main files that i didnt need added multiple colour options for a bit of variety. made all of the Collision located within the GameObject class. for ease of access. made finger input an array rather then an arraylist. added a lot of paints to the global class, to hopefully reduce paint creation every time i create a similar object. added a boomerang object type so that it has unique interactions with spells as always ive probably forgotten lots of things. tried solving link's bugout problems. inconclusive

+++ trunk/WarlocksGradle/MyApplication/src/main/java/Tools/BoundingCircle.java
@@ -39,8 +39,8 @@ public class BoundingCircle {
         float cx = 0;
         float cy = 0;
         if (det != 0) {
-            cx = (float) ((A1 * C1 - B1 * C2) / det);
-            cy = (float) ((A1 * C2 - -B1 * C1) / det);
+            cx = (A1 * C1 - B1 * C2) / det;
+            cy = (A1 * C2 - -B1 * C1) / det;
         } else {
             cx = x0;
             cy = y0;
@@ -49,7 +49,7 @@ public class BoundingCircle {
     }
 
     public void Draw(Canvas c, float playerx, float playery, Paint p) {
-        c.drawCircle(Center.x - playerx, Center.y - playery, (float) Radius, p);
+        c.drawCircle(Center.x - playerx, Center.y - playery, Radius, p);
     }
 
 }
